+++ packages/auth-lti/isomorphic/constants.js
@@ -1,0 +1,2 @@
+export const LOGIN_URL = '/auth/lti'
+export const CALLBACK_URL = '/auth/lti/callback'